+++ invenio_rdm_records/__init__.py
@@ -3,6 +3,7 @@
 # Copyright (C) 2019-2024 CERN.
 # Copyright (C) 2019-2024 Northwestern University.
 # Copyright (C) 2024      KTH Royal Institute of Technology.
+# Copyright (C) 2024 Graz University of Technology.
 #
 # Invenio-RDM-Records is free software; you can redistribute it and/or modify
 # it under the terms of the MIT License; see LICENSE file for more details.
@@ -11,6 +12,6 @@
 
 from .ext import InvenioRDMRecords
 
-__version__ = "15.6.0"
+__version__ = "16.0.1"
 
 __all__ = ("__version__", "InvenioRDMRecords")