+++ tests/requests/conftest.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2019-2021 CERN.
+# Copyright (C) 2019-2024 CERN.
 # Copyright (C) 2019-2021 Northwestern University.
 #
 # Invenio-RDM-Records is free software; you can redistribute it and/or modify
@@ -15,6 +15,8 @@ fixtures are available.
 import pytest
 from invenio_records_permissions.generators import AuthenticatedUser, SystemProcess
 
+import invenio_rdm_records.services.communities.moderation as communities_moderation
+from invenio_rdm_records.services.components.verified import UserModerationHandler
 from invenio_rdm_records.services.permissions import RDMRecordPermissionPolicy
 
 
@@ -35,4 +37,9 @@ def app_config(app_config):
     app_config["RDM_PERMISSION_POLICY"] = CustomRDMRecordPermissionPolicy
     # Enable user moderation
     app_config["RDM_USER_MODERATION_ENABLED"] = True
+    # Enable content moderation handlers
+    app_config["RDM_CONTENT_MODERATION_HANDLERS"] = [UserModerationHandler()]
+    app_config["RDM_COMMUNITY_CONTENT_MODERATION_HANDLERS"] = [
+        communities_moderation.UserModerationHandler(),
+    ]
     return app_config